Add SpEED Matlab FE and QR and move them with ST-RRED FE and QR to new files, add unittests, move testutil in tools.
