+++ python/src/vmaf/core/quality_runner.py
@@ -13,7 +13,7 @@ from vmaf.core.feature_assembler import FeatureAssembler
 from vmaf.core.train_test_model import TrainTestModel, LibsvmNusvrTrainTestModel, \
     BootstrapLibsvmNusvrTrainTestModel
 from vmaf.core.feature_extractor import SsimFeatureExtractor, MsSsimFeatureExtractor, \
-    VmafFeatureExtractor, StrredFeatureExtractor
+    VmafFeatureExtractor
 
 __copyright__ = "Copyright 2016-2018, Netflix, Inc."
 __license__ = "Apache, Version 2.0"
@@ -893,57 +893,6 @@ class MotionQualityRunner(VmafSingleFeatureQualityRunner):
     TYPE = 'MOTION'
     # TYPE = 'TI'
     FEATURE_NAME = 'motion'
-
-class StrredQualityRunner(QualityRunner):
-
-    TYPE = 'STRRED'
-
-    # VERSION = '1.0'
-    VERSION = 'F' + StrredFeatureExtractor.VERSION + '-1.1'
-
-    def _get_quality_scores(self, asset):
-        raise NotImplementedError
-
-    def _generate_result(self, asset):
-        raise NotImplementedError
-
-    def _get_feature_assembler_instance(self, asset):
-
-        feature_dict = {StrredFeatureExtractor.TYPE: StrredFeatureExtractor.ATOM_FEATURES + getattr(StrredFeatureExtractor, 'DERIVED_ATOM_FEATURES', [])}
-
-        feature_assembler = FeatureAssembler(
-            feature_dict=feature_dict,
-            feature_option_dict=None,
-            assets=[asset],
-            logger=self.logger,
-            fifo_mode=self.fifo_mode,
-            delete_workdir=self.delete_workdir,
-            result_store=self.result_store,
-            optional_dict=None,
-            optional_dict2=None,
-            parallelize=False, # parallelization already in a higher level
-        )
-        return feature_assembler
-
-    def _run_on_asset(self, asset):
-        # Override Executor._run_on_asset(self, asset)
-        vmaf_fassembler = self._get_feature_assembler_instance(asset)
-        vmaf_fassembler.run()
-        feature_result = vmaf_fassembler.results[0]
-        result_dict = {}
-        result_dict.update(feature_result.result_dict.copy()) # add feature result
-        result_dict[self.get_scores_key()] = feature_result.result_dict[
-            StrredFeatureExtractor.get_scores_key('strred')] # add strred score
-        del result_dict[StrredFeatureExtractor.get_scores_key('strred')] # delete redundant
-        return Result(asset, self.executor_id, result_dict)
-
-    def _remove_result(self, asset):
-        # Override Executor._remove_result(self, asset) by redirecting it to the
-        # FeatureAssembler.
-
-        vmaf_fassembler = self._get_feature_assembler_instance(asset)
-        vmaf_fassembler.remove_results()
-
 
 class BootstrapVmafQualityRunner(VmafQualityRunner):
