+++ python/test/feature_assembler_test.py
@@ -6,7 +6,7 @@ import unittest
 from vmaf.core.feature_assembler import FeatureAssembler
 from vmaf.core.feature_extractor import VmafFeatureExtractor, FeatureExtractor, \
     MomentFeatureExtractor
-from testutil import set_default_576_324_videos_for_testing
+from vmaf.tools.testutil import set_default_576_324_videos_for_testing
 
 
 class FeatureAssemblerTest(unittest.TestCase):
