+++ python/test/result_test.py
@@ -1,4 +1,4 @@
-from testutil import set_default_576_324_videos_for_testing
+from vmaf.tools.testutil import set_default_576_324_videos_for_testing
 
 __copyright__ = "Copyright 2016-2018, Netflix, Inc."
 __license__ = "Apache, Version 2.0"
